+++ src/world/world_render.c
@@ -60,14 +60,14 @@ static void chunk_lod_render(chunk_lod_t *lod)
 
 static void chunk_render(chunk_t *chunk)
 {
-    float dist = vec2_dist(
-    (vec2){chunk->pos.x * CHUNK_SIZE + CHUNK_SIZE / 2.0f,
-    chunk->pos.y * CHUNK_SIZE + CHUNK_SIZE / 2.0f},
+    double dist = vec2_dist(
+    (vec2){chunk->pos.x * CHUNK_SIZE + CHUNK_SIZE / 2.0,
+    chunk->pos.y * CHUNK_SIZE + CHUNK_SIZE / 2.0},
     (vec2){_demo->cam.pos.x, _demo->cam.pos.z});
 
-    if (dist < 1500.0f)
+    if (dist < 1500.0)
         chunk_lod_render(&chunk->lod[2]);
-    else if (dist < 3000.0f)
+    else if (dist < 3000.0)
         chunk_lod_render(&chunk->lod[1]);
     else
         chunk_lod_render(&chunk->lod[0]);
